Test: Order Test 구현
- Service의 Test코드 구현

+++ src/main/java/run/bemin/api/order/dto/response/PagesResponse.java
@@ -11,5 +11,5 @@ public class PagesResponse<T> {
   private int pageNumber; // 현재 페이지 번호
   private int pageSize; // 페이지 크기
   private int totalPages; // 전체 페이지 수
-  private int totalElements; // 전체 데이터 수
+  private Long totalElements; // 전체 데이터 수
 }
